+++ Rocket.Chat.Shared/Managers/Launcher/LauncherProtocol.swift
@@ -1,0 +1,23 @@
+//
+//  LauncherProtocol.swift
+//  Rocket.Chat
+//
+//  Created by Lucas Woo on 6/17/17.
+//  Copyright © 2017 Rocket.Chat. All rights reserved.
+//
+
+import UIKit
+
+protocol LauncherProtocol {
+    var coordinators: [LauncherCoordinator] { get }
+}
+
+extension LauncherProtocol {
+    func prepareToLaunch(with options: [UIApplicationLaunchOptionsKey: Any]?) {
+        coordinators.forEach { $0.prepareToLaunch(with: options) }
+    }
+}
+
+protocol LauncherCoordinator {
+    func prepareToLaunch(with options: [UIApplicationLaunchOptionsKey: Any]?)
+}
