--- Rocket.Chat/Managers/Launcher/LauncherProtocol.swift
@@ -1,27 +1,0 @@
-//
-//  LauncherProtocol.swift
-//  Rocket.Chat
-//
-//  Created by Rafael Machado on 11/12/16.
-//  Copyright © 2016 Rocket.Chat. All rights reserved.
-//
-
-import UIKit
-
-protocol LauncherProtocol {
-    func prepareToLaunch(with options: [UIApplicationLaunchOptionsKey: Any]?)
-}
-
-final class Launcher: LauncherProtocol {
-    private lazy var launchers: [LauncherProtocol] = {
-        return [
-            PersistencyCoordinator(),
-            BugTrackingCoordinator(),
-            UserCoordinator()
-        ]
-    }()
-
-    func prepareToLaunch(with options: [UIApplicationLaunchOptionsKey: Any]?) {
-        launchers.forEach { $0.prepareToLaunch(with: options) }
-    }
-}
